Fix compiler warnings
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/branches/bug_fix@69248 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/lib/window3D/orientationAnnotation.cc
@@ -231,7 +231,7 @@ Point3df OrientationAnnotation::getPosition(
   Point3df coords = q.transform( Point3df( 0, 0, 0 ) );
 
   // Compute the coordinates (without shift adjustment)
-  for ( int i = 0 ; i < _winCoordParamIndexes[ pos ].size() ; i++ )
+  for ( size_t i = 0 ; i < _winCoordParamIndexes[ pos ].size() ; i++ )
   {
     coords[ i ] += _winCoordParams [ _winCoordParamIndexes[ pos ][ i ] ][ i ];
     if ( _winCoordParamIndexes[ pos ][ i ] == 3 )
